+++ test/Spec.hs
@@ -18,6 +18,7 @@ import qualified Canon.Git.ParseTest
 import qualified Canon.Model.YamlTest
 import qualified Canon.RegistryTest
 import qualified Canon.VettingTest
+import qualified Canon.TestingTest
 import qualified CanonTest
 import Test.Tasty (defaultMain, testGroup)
 
@@ -35,7 +36,7 @@ main =
             , Canon.Antlr4.QueryTest.tests
             , Canon.Antlr4.RuleGraphTest.tests
             , Canon.Model.YamlTest.tests
-            , Canon.RegistryTest.tests, Canon.VettingTest.tests
+            , Canon.RegistryTest.tests, Canon.VettingTest.tests, Canon.TestingTest.tests
             , Canon.Git.ParseTest.tests
             , Canon.CanonicalCommentTest.tests
             , Canon.AttachTest.tests